+++ docs/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 0.2.0 (under development)
-Remove verified icon (use https://marketplace.humhub.com/module/verified/description instead)
+- Enh: make icon color configurable
+- Enh: use color and icon pickers
+- Remove default values for configuration and drop support for file configuration
+- Remove verified icon (use https://marketplace.humhub.com/module/verified/description instead)
 
 ## 0.1.0 (unreleased)
 First release